Consolidate duplicate InbuiltFunctionEvaluatorTest classes
